+++ src/Products/PloneMeeting/utils.py
@@ -1,24 +1,4 @@
 # -*- coding: utf-8 -*-
-#
-# Copyright (c) 2015 by Imio.be
-#
-# GNU General Public License (GPL)
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-# 02110-1301, USA.
-#
 
 from AccessControl.Permission import Permission
 from appy.shared.diff import HtmlDiff
@@ -1666,6 +1646,11 @@ def decodeDelayAwareId(delayAwareId):
        'orgauid__rowid__myuniquerowid.20141215'. We return the org_uid and the row_id."""
     infos = delayAwareId.split('__rowid__')
     return infos[0], infos[1]
+
+
+def uncapitalize(string):
+    """Lowerize first letter of given p_string."""
+    return string[0].lower() + string[1:]
 
 
 class AdvicesUpdatedEvent(ObjectEvent):